requirements.txt
yt-dlp
ffmpeg-python
imageio-ffmpeg
Pillow
requests
